rename : 컴포넌트내의 관련 폴더 생성후 파일 이동
