+++ web/src/components/pay-link/PayLinkGuideCarousel.tsx
